+++ src/components/content/content-filter-panel.tsx
@@ -40,7 +40,7 @@ function ContentFilterPanel<T extends string>({
             aria-label="Search"
             variant={"outline"}
             size={"lg"}
-            bgColor={{ base: "bg", _hover: "bg.muted" }}
+            bgColor={{ base: "bg.muted", _hover: "bg" }}
             borderRadius={"md"}
           >
             <FontAwesomeIcon icon={faSearch} />